[Android SDK] Provide test helpers (#19580)
* Add a test with mock

* Add empty core testing support module

* Use fine grained optin

Because optin was setup for module wide, it was breaking modules that did not have the optin in their classpath. For fixing that, start using fine grain optin. It should even be a good practice as we will be able to track more precisely experimental API used in the SDK.

* Make first TimelineBuilder test

* Backport all suitable tests for TimelineBuilder

* Run tests from message-ui module on CI

* Add audio diffutil tests

* Add conversation diffutil tests

* Add image diffutil tests

* Split fakers

* Dump api

* Generate jenkins files

* Rename test module