Use new ktlint api (#26244)
- Related to: #26199 

## Summary

Updates our ktlint jar to use the new ktlint api.
Here's their changelog:
https://github.com/pinterest/ktlint/blob/master/CHANGELOG.md


### Changelog

- visit became **beforeVisitChildNodes**
[docs](https://github.com/pinterest/ktlint/blob/master/CHANGELOG.md#rule-lifecycle-hooks--deprecate-runonrootonly-visitor-modifier)
`
beforeVisitChildNodes: This method is called on a node in AST before
visiting its child nodes. This is repeated recursively for the child
nodes resulting in a depth first traversal of the AST. This method is
the equivalent of the "visit" life cycle hooks. However, note that in
KtLint 0.48, the UserData of the rootnode no longer provides access to
the ".editorconfig" properties. This method can be used to emit Lint
Violations and to autocorrect if applicable.
`

- Trailing comma is default
[docs](https://github.com/pinterest/ktlint/blob/master/CHANGELOG.md#changed-3)

- Use RuleSetProviderV2
[docs](https://github.com/pinterest/ktlint/blob/master/CHANGELOG.md#ruleset-providing-by-custom-rule-set-provider)

- A LOT OF FORMATTING

---------

Co-authored-by: Arthur Lacoste <arthur@nabla.com>